requirements.txt
numpy
pandas
torch
scikit-learn
emoji
tweet-preprocessor
nltk
contractions
transformers
datasets
evaluate
accelerate
